% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expand_daily}
\alias{expand_daily}
\title{Expand xts do daily periodicity.}
\usage{
expand_daily(x, enddate = "today")
}
\arguments{
\item{x}{xts object}

\item{enddate}{last date.}
}
\value{
xts
}
\description{
Expand xts do daily periodicity.
}
\examples{

x = get_nbp_interest_rates()
expand_daily(x)

}
