% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_max_loan}
\alias{get_max_loan}
\title{Maximum interest rates for loans in Poland}
\usage{
get_max_loan()
}
\value{
RETURN DESCRIPTION
}
\description{
Get time series with the maximum interest rates for loans in Poland.
}
\examples{

get_max_loan()

}

