% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_nbp_interest_rates}
\alias{get_nbp_interest_rates}
\title{Download interest rates from National Bank of Poland.}
\usage{
get_nbp_interest_rates(types = c("ref", "lom", "dep", "red"))
}
\arguments{
\item{types}{DESCRIPTION.}
}
\value{
xts object.
}
\description{
This functions download various interest rates from NBP (National Bank of Poland) as a xts object.
}
\examples{

tail(get_nbp_interest_rates("ref"))
tail(get_nbp_interest_rates(c("ref","lom")))

}
