% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.R
\name{nbp_inflation_download}
\alias{nbp_inflation_download}
\title{Download inflation data from NBP website.}
\usage{
nbp_inflation_download(destination = sprintf("inflation_nbp-\%s.xls",
  Sys.Date()), overwrite = FALSE, ...)
}
\arguments{
\item{destination}{A character string with the name where the downloaded file is saved. Tilde-expansion is performed.}

\item{overwrite}{if true overwrites the file in the destination path.}
}
\description{
Download inflation data from NBP website.
}
\examples{

path <- nbp_inflation_download()

}
