% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nbp_xts2tbl}
\alias{nbp_xts2tbl}
\title{Convert xts to tibble.}
\usage{
nbp_xts2tbl(x)
}
\arguments{
\item{x}{xts object.}
}
\value{
tbl_df object.
}
\description{
Adds column with dates to the first position to after conversion to tibble.
}
\examples{

x = nbp_interest_rates()
tail(nbp_xts2tbl(x))

dx = nbp_expand_daily(x)
tail(nbp_xts2tbl(dx))


}
