% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xts2tbl}
\alias{xts2tbl}
\title{Convert xts to tibble.}
\usage{
xts2tbl(x)
}
\arguments{
\item{x}{xts object.}
}
\value{
tbl_df object.
}
\description{
Adds column with dates to the first position to after conversion to tibble.
}
\examples{

x = get_interest_rates()
xts2tbl(x)

dx = expand_daily(x)
xts2tbl(dx)


}

